Move the MockFileObject to its own file.
