+++ tests/logic_layer_t/test_create_new_attachment.py
@@ -7,16 +7,8 @@ from werkzeug.exceptions import NotFound, Forbidden
 from models.attachment import Attachment
 from models.task import Task
 from models.user import User
+from tests.util import MockFileObject
 from util import generate_ll
-
-
-class MockFileObject(object):
-    def __init__(self, filename):
-        self.filename = filename
-        self.save_calls = []
-
-    def save(self, filepath):
-        self.save_calls.append(filepath)
 
 
 class CreateNewAttachmentTest(unittest.TestCase):
